Tried out naked roy
Trying out naked roy to see where the issues of pinching were coming from. It looks like its the model itself. I am deciding on what to do and would be the easiest to accomplish. Due to time constraints it be easier to go into each animation and hand clean it instead of having the model redone since it took several months to complete.