Fixed Level trans to cath, building bug
Cathedral level transition was a problem with game mode begin UI and not UI/Game. Building bug was due to changes to a parent class whose child was not upgraded to the new version. However it is no longer being used as far as I could tell so the code was gutted to make the error go away.