Think I figured out the problem with heavy mook
in the heavy mook tree you created i went into the attack and it still had the my light mook instead of my heavy mook instead. also with the play animation it was still calling the light mook so i switched it to heavy mook montage so lets see if that might work.